requirements.txt
watchdog
python-magic-bin
customtkinter
pystray
pillow